Revert "Added udmabuf and demo_dma to user level test bench. I still need too add irqs"
This reverts commit 4cedde402e48d03f5e68ad27bfe96fbb601db9ec.

+++ systemc/memory.h
@@ -46,7 +46,7 @@ public:
 	void accelerate();
 
 	processor_tb *processor_test_bench;
-	uint8_t	*mem;	
+	float	*mem;	
 
 	sc_event processor_start;
 	SC_HAS_PROCESS(memory);
